moved the speed computation method to the base particle structure
